move classes to correct namespaces
