+++ src/CrawlProfiles/CrawlInternalUrls.php
@@ -1,9 +1,10 @@
 <?php
 
-namespace Spatie\Crawler;
+namespace Spatie\Crawler\CrawlProfiles;
 
 use GuzzleHttp\Psr7\Uri;
 use Psr\Http\Message\UriInterface;
+use Spatie\Crawler\CrawlProfiles\CrawlProfile;
 
 class CrawlInternalUrls extends CrawlProfile
 {
