+++ src/CrawlQueues/ArrayCrawlQueue.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Spatie\Crawler\CrawlQueue;
+namespace Spatie\Crawler\CrawlQueues;
 
 use Psr\Http\Message\UriInterface;
 use Spatie\Crawler\CrawlUrl;
